PKLUV-3168 Bruker `@PathParam` i `startEskalering` i stedet for json

+++ common/src/main/java/no/nav/fo/veilarbsituasjon/db/SituasjonRepository.java
@@ -344,7 +344,7 @@ public class SituasjonRepository {
     }
 
     @Transactional
-    public void startEskalering(String aktorId, String opprettetAv, int tilhorendeDialogId) {
+    public void startEskalering(String aktorId, String opprettetAv, long tilhorendeDialogId) {
         if (hentEskaleringstatus(aktorId) != null) {
             throw new RuntimeException();
         }
